% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{autoplot.daa_performance}
\alias{autoplot.daa_performance}
\title{Plot Method for DAA Performance Objects}
\usage{
\method{autoplot}{daa_performance}(object, ...)
}
\arguments{
\item{object}{A daa_performance object containing performance metrics}

\item{...}{Additional arguments passed to plot_performance}
}
\value{
A ggplot2 object
}
\description{
Automatically creates a visualization of DAA performance results using ggplot2.
}
