% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{calculate_overlap_ratio}
\alias{calculate_overlap_ratio}
\title{Calculate overlap ratio between two sets of significant features}
\usage{
calculate_overlap_ratio(set1, set2)
}
\arguments{
\item{set1}{First set of feature names}

\item{set2}{Second set of feature names}
}
\value{
Numeric value between 0 and 1
}
\description{
Calculate overlap ratio between two sets of significant features
}
