% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_daa_methods.R
\name{create_summary}
\alias{create_summary}
\title{Create Summary of DAA Results}
\usage{
create_summary(results, methods)
}
\arguments{
\item{results}{List of results from different DAA methods}

\item{methods}{Character vector of method names}
}
\value{
A data frame with columns:
\itemize{
\item Method: Name of the DAA method
\item N_Significant: Number of features called as significant
\item Mean_Effect: Mean absolute effect size
}
}
\description{
Create Summary of DAA Results
}
\keyword{internal}
