% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{get_test_statistics}
\alias{get_test_statistics}
\title{Get test statistics from DAA results}
\usage{
get_test_statistics(results)
}
\arguments{
\item{results}{DAA analysis results}
}
\value{
Numeric vector of test statistics
}
\description{
Get test statistics from DAA results
}
