% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_from_biom}
\alias{import_from_biom}
\title{Import Data from BIOM Format}
\usage{
import_from_biom(file_path, ...)
}
\arguments{
\item{file_path}{Character string specifying the path to the BIOM file}

\item{...}{Additional arguments passed to the import function}
}
\value{
A phyloseq object containing the imported data
}
\description{
Import Data from BIOM Format
}
\examples{
\dontrun{
data <- import_from_biom("path/to/file.biom")
}

}
\keyword{internal}
