% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_from_csv}
\alias{import_from_csv}
\title{Import Data from CSV Format}
\usage{
import_from_csv(file_path, taxonomy_cols = NULL, ...)
}
\arguments{
\item{file_path}{Character string specifying the path to the CSV file}

\item{taxonomy_cols}{Character vector specifying which columns contain taxonomy information}

\item{...}{Additional arguments passed to read.csv}
}
\value{
A phyloseq object containing the imported data
}
\description{
Import Data from CSV Format
}
\examples{
\dontrun{
data <- import_from_csv("path/to/file.csv", 
                       taxonomy_cols = c("Kingdom", "Phylum", "Class"))
}

}
\keyword{internal}
