% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_from_phyloseq}
\alias{import_from_phyloseq}
\title{Import Data from Phyloseq Format}
\usage{
import_from_phyloseq(file_path)
}
\arguments{
\item{file_path}{Character string specifying the path to the saved phyloseq object}
}
\value{
A phyloseq object containing the imported data
}
\description{
Import Data from Phyloseq Format
}
\examples{
\dontrun{
data <- import_from_phyloseq("path/to/saved_phyloseq.RData")
}

}
\keyword{internal}
