% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{print.daa_comparison}
\alias{print.daa_comparison}
\title{Print Method for DAA Comparison Objects}
\usage{
\method{print}{daa_comparison}(x, ...)
}
\arguments{
\item{x}{A daa_comparison object to print}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for DAA Comparison Objects
}
\examples{
\dontrun{
comparison_results <- compare_methods(performance_data)
print(comparison_results)
}

}
