% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{print.daa_performance}
\alias{print.daa_performance}
\title{Print method for daa_performance objects}
\usage{
\method{print}{daa_performance}(x, ...)
}
\arguments{
\item{x}{daa_performance object}

\item{...}{Additional arguments}
}
\description{
Print method for daa_performance objects
}
