% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{print.daa_performance_plot}
\alias{print.daa_performance_plot}
\title{Print Method for DAA Performance Plot Objects}
\usage{
\method{print}{daa_performance_plot}(x, ...)
}
\arguments{
\item{x}{A daa_performance_plot object to print}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for DAA Performance Plot Objects
}
\examples{
\dontrun{
perf_plot <- plot_performance(performance_data)
print(perf_plot)
}

}
