% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{plot.daa_performance}
\alias{plot.daa_performance}
\title{Plot method for daa_performance objects}
\usage{
\method{plot}{daa_performance}(x, metric = "sensitivity", ...)
}
\arguments{
\item{x}{A daa_performance object}

\item{metric}{Performance metric to plot ("sensitivity", "specificity", "precision", "f1_score")}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for daa_performance objects
}
